+++ LanAnhComputer.API/Data/Entities/ProductImage.cs
@@ -2,8 +2,8 @@ namespace LanAnhComputer.Data.Entities
 {
     public class ProductImage
     {
-        public int ProductImageId { get; set; }
-        public int ProductId { get; set; }
+        public long ProductImageId { get; set; }
+        public long ProductId { get; set; }
         public string ImageUrl { get; set; } = null!;
         public string? AltText { get; set; }
         public bool IsPrimary { get; set; }
