+++ LanAnhComputer.API/Data/Entities/ProductSpecification.cs
@@ -2,8 +2,8 @@ namespace LanAnhComputer.Data.Entities
 {
     public class ProductSpecification
     {
-        public int ProductSpecificationId { get; set; }
-        public int ProductId { get; set; }
+        public long ProductSpecificationId { get; set; }
+        public long ProductId { get; set; }
         public string GroupName { get; set; } = "General";
         public string SpecKey { get; set; } = null!;
         public string SpecValue { get; set; } = null!;
